requirements.txt
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-Admin==1.4.2
flask-paginate==0.4.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
pbr==1.10.0
python-slugify==1.2.1
six==1.10.0
SQLAlchemy==1.1.4
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Unidecode==0.4.21
Werkzeug==0.11.11
WTForms==2.1
